requirements.txt
aiogram==2.19
aiohttp==3.8.1
aiosignal==1.2.0
aiosqlite==0.17.0
alembic==1.7.7
async-timeout==4.0.2
asyncpg==0.25.0
attrs==21.4.0
Babel==2.9.1
certifi==2021.10.8
charset-normalizer==2.0.12
databases==0.5.5
frozenlist==1.3.0
greenlet==1.1.2
idna==3.3
loguru==0.6.0
Mako==1.2.0
MarkupSafe==2.1.1
multidict==6.0.2
ormar==0.10.25
psycopg2-binary==2.9.3
pydantic==1.9.0
pytz==2022.1
PyYAML==6.0
SQLAlchemy==1.4.31
typing-extensions==4.1.1
yarl==1.7.2
